+++ src/main/java/fr/siamois/bean/SessionSettings.java
@@ -1,5 +1,6 @@
 package fr.siamois.bean;
 
+import fr.siamois.models.Institution;
 import fr.siamois.models.Team;
 import fr.siamois.models.auth.Person;
 import fr.siamois.models.exceptions.NoTeamSelectedException;
@@ -17,11 +18,14 @@ import javax.faces.bean.SessionScoped;
 public class SessionSettings {
 
     private Team selectedTeam;
+    @Getter
+    private Institution selectedInstitution;
 
     public Person getAuthenticatedUser() {
         return AuthenticatedUserUtils.getAuthenticatedUser().orElseThrow(() -> new RuntimeException("No authenticated user"));
     }
 
+    @Deprecated
     public Team getSelectedTeam() {
         if (selectedTeam == null) {
             throw new NoTeamSelectedException();
