+++ src/main/java/fr/siamois/models/Institution.java
@@ -4,8 +4,6 @@ import fr.siamois.models.auth.Person;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
 @Data
 @Entity
