+++ src/main/java/fr/siamois/models/TraceInfo.java
@@ -1,0 +1,5 @@
+package fr.siamois.models;
+
+import fr.siamois.models.auth.Person;
+
+public record TraceInfo(Institution institution, Person author) {}
