+++ src/main/java/fr/siamois/services/HistoryService.java
@@ -1,9 +1,12 @@
 package fr.siamois.services;
 
 import fr.siamois.infrastructure.repositories.history.GlobalHistoryRepository;
+import fr.siamois.infrastructure.repositories.history.GlobalHistoryTeamRepository;
 import fr.siamois.infrastructure.repositories.history.SpatialUnitHistoryRepository;
+import fr.siamois.models.Institution;
 import fr.siamois.models.SpatialUnit;
 import fr.siamois.models.Team;
+import fr.siamois.models.TraceInfo;
 import fr.siamois.models.auth.Person;
 import fr.siamois.models.history.HistoryOperation;
 import fr.siamois.models.history.HistoryUpdateType;
@@ -27,31 +30,31 @@ public class HistoryService {
     private static final List<String> tableNames = List.of("action_unit", "recording_unit",
             "recording_unit_study", "siamois_document", "spatial_unit", "specimen", "specimen_study");
 
-    public HistoryService(SpatialUnitHistoryRepository spatialUnitHistoryRepository, GlobalHistoryRepository globalHistoryRepository) {
+    public HistoryService(SpatialUnitHistoryRepository spatialUnitHistoryRepository, GlobalHistoryTeamRepository globalHistoryRepository) {
         this.spatialUnitHistoryRepository = spatialUnitHistoryRepository;
         this.globalHistoryRepository = globalHistoryRepository;
     }
 
 
-    public List<HistoryOperation> findAllOperationsOfUserAndTeamBetween(Person person, Team team, OffsetDateTime start, OffsetDateTime end) {
+    public List<HistoryOperation> findAllOperationsOfUserAndTeamBetween(TraceInfo info, OffsetDateTime start, OffsetDateTime end) {
         if (start.isAfter(end)) return new ArrayList<>();
 
         List<HistoryOperation> operations = new ArrayList<>();
-        addAllCreationOperations(person, team, operations, start, end);
-        addAllHistoryOperations(person, team, operations, start, end);
+        addAllCreationOperations(info, operations, start, end);
+        addAllHistoryOperations(info, operations, start, end);
 
         operations.sort(Comparator.comparing(HistoryOperation::actionDatetime).reversed());
 
         return operations;
     }
 
-    private void addAllHistoryOperations(Person person, Team team, List<HistoryOperation> operations, OffsetDateTime beginTime, OffsetDateTime endTime) {
+    private void addAllHistoryOperations(TraceInfo info, List<HistoryOperation> operations, OffsetDateTime beginTime, OffsetDateTime endTime) {
         for (int i = 0; i < entityName.size(); i++)
-            addHistoryOperation(operations, tableNames.get(i), entityName.get(i), person, team, beginTime, endTime);
+            addHistoryOperation(operations, tableNames.get(i), entityName.get(i), info, beginTime, endTime);
     }
 
-    private void addHistoryOperation(List<HistoryOperation> operations, String tableName, String entityName, Person person, Team team, OffsetDateTime start, OffsetDateTime end) {
-        globalHistoryRepository.findAllHistoryOfUserBetween("history_" + tableName, person, team, start, end).forEach((entry) ->
+    private void addHistoryOperation(List<HistoryOperation> operations, String tableName, String entityName, TraceInfo info, OffsetDateTime start, OffsetDateTime end) {
+        globalHistoryRepository.findAllHistoryOfUserBetween("history_" + tableName, info, start, end).forEach((entry) ->
                 operations.add(new HistoryOperation(entry.getUpdateType(),
                         entityName,
                         entry.getTableId(),
@@ -59,8 +62,8 @@ public class HistoryService {
                 );
     }
 
-    private void addCreateOperation(List<HistoryOperation> operations, String tableName, String entityName, Person person, Team team, OffsetDateTime start, OffsetDateTime end) {
-        globalHistoryRepository.findAllCreationOfUserBetween(tableName, person, team,  start, end).forEach((entity) ->
+    private void addCreateOperation(List<HistoryOperation> operations, String tableName, String entityName, TraceInfo info, OffsetDateTime start, OffsetDateTime end) {
+        globalHistoryRepository.findAllCreationOfUserBetween(tableName, info,  start, end).forEach((entity) ->
                 operations.add(new HistoryOperation(HistoryUpdateType.CREATE,
                         entityName,
                         entity.getId(),
@@ -68,9 +71,9 @@ public class HistoryService {
                 );
     }
 
-    private void addAllCreationOperations(Person person, Team team, List<HistoryOperation> operations, OffsetDateTime beginTime, OffsetDateTime endTime) {
+    private void addAllCreationOperations(TraceInfo info, List<HistoryOperation> operations, OffsetDateTime beginTime, OffsetDateTime endTime) {
         for (int i = 0; i < entityName.size(); i++)
-            addCreateOperation(operations, tableNames.get(i), entityName.get(i), person, team, beginTime, endTime);
+            addCreateOperation(operations, tableNames.get(i), entityName.get(i), info, beginTime, endTime);
     }
 
     public List<SpatialUnitHist> findSpatialUnitHistory(SpatialUnit current) {
